Create a methylTracer object from an HDF5 file

Construct a methylTracer object from an HDF5 file that stores methylation data and associated metadata. The HDF5 file is expected to contain a main assay dataset "X" and accompanying sample and marker identifiers organised in the "/obs" and "/var" groups, respectively.

build_met_obj(
  h5_file = NULL,
  sample_name = "sample_name",
  marker_name = "marker_name"
)

Arguments

h5_file

Character string giving the path to the HDF5 file containing methylation data and metadata.

sample_name

Character scalar. Name of the dataset inside the "/obs" group that contains sample identifiers (default: "sample_name"). The corresponding HDF5 path is "/obs/<sample_name>".

marker_name

Character scalar. Name of the dataset inside the "/var" group that contains marker identifiers (default: "marker_name"). The corresponding HDF5 path is "/var/<marker_name>".

Value

A methylTracer object with:

the main assay backed by an HDF5Array pointing to "X" in h5_file;

marker_name: an HDF5Array vector pointing to "/var/<marker_name>";

sample_name: an HDF5Array vector pointing to "/obs/<sample_name>".

If an error occurs (e.g. missing datasets), a diagnostic message is printed and NULL is returned.

Details

The typical layout of an HDF5 file for methylTracer is:

"X": a 2D dataset (rows = markers/regions, cols = samples), storing methylation values as proportions in \([0, 1]\).

"/obs/<sample_name>": a 1D dataset containing sample IDs.

"/var/<marker_name>": a 1D dataset containing marker IDs, usually of the form "chr_start_end".

This function wraps the main assay "X" and the specified sample_name and marker_name datasets into a methylTracer object using HDF5Array for on-disk storage.

Examples

suppressMessages(library(HDF5Array))
toy_h5 <- system.file("extdata", "toy.h5", package = "HDF5Array")
X <- HDF5Array::HDF5Array(toy_h5, "M1")

num_rows <- dim(X)[1]
start_values <- seq(from = 10000, by = 1000, length.out = num_rows)
end_values <- start_values + 1000
marker_name <- paste0("chr1_", start_values, "_", end_values)
sample_name <- paste0("sample_", seq_len(dim(X)[2]))
group <- c(rep("A", 50), rep("B", 50), rep("C", 50))
regul <- c(rep("exon", 50), rep("intron", 50), rep("TSS", 50))

current_dir <- tempdir()
met_obj_test <- file.path(current_dir, "methylTracer_obj_test.h5")
unlink(met_obj_test)  # Remove existing file if it exists

rhdf5::h5createFile(met_obj_test)
rhdf5::h5createGroup(met_obj_test, "obs")
rhdf5::h5createGroup(met_obj_test, "var")

HDF5Array::writeHDF5Array(X, met_obj_test, "X")
#> <10000 x 150> HDF5Matrix object of type "double":
#>                [,1]       [,2]       [,3] ...     [,149]     [,150]
#>     [1,] 0.44015009 0.02303051 0.11688992   .  0.6684808  0.9859835
#>     [2,] 0.18883376 0.84313244 0.12153315   .  0.2886549  0.2750908
#>     [3,] 0.99232698 0.22924928 0.20998143   .  0.4786133  0.7087926
#>     [4,] 0.52631346 0.15438076 0.43283016   .  0.8715839  0.5870442
#>     [5,] 0.62369062 0.27622849 0.54027815   .  0.3082406  0.2929455
#>      ...          .          .          .   .          .          .
#>  [9996,] 0.69183231 0.49322320 0.69210429   . 0.50506635 0.17500671
#>  [9997,] 0.01577389 0.64192487 0.70604724   . 0.80835509 0.09372537
#>  [9998,] 0.51859423 0.78645107 0.51546450   . 0.76275510 0.18370904
#>  [9999,] 0.29122618 0.21718068 0.54856003   . 0.14534205 0.97475079
#> [10000,] 0.87539984 0.34665351 0.06661729   . 0.71398136 0.22140136
rhdf5::h5write(marker_name, met_obj_test, "var/marker_name")
rhdf5::h5write(regul, met_obj_test, "var/regul")
rhdf5::h5write(sample_name, met_obj_test, "obs/sample_name")
rhdf5::h5write(group, met_obj_test, "obs/group")

met_obj_test <- build_met_obj(
  met_obj_test,
  sample_name = "sample_name",
  marker_name = "marker_name"
)

met_obj_test
#> methylTracer Object Summary:
#> Local_path    : /tmp/RtmpRfaB9m/methylTracer_obj_test.h5
#> Memory use    : 7.856 KB
#> Dimensions:
#> X   dims  : 10000 x 150
#> var dims  : 10000 x 2
#> obs dims  : 150 x 2