Compute Column Means for Variable Groups

This function computes the mean of selected columns (variables) from an HDF5 dataset based on specified group indices in the regulation column (var_group). It retrieves data from an HDF5 file, calculates the column-wise means, and writes the result to a new dataset in the same HDF5 file. The result is saved under a variable name combining the dataset and regulation group names.

computeVar(met, dataset = "/X", varColName = NULL)

Arguments

met

A methylTracer object containing metadata.

dataset

A string specifying the dataset path inside the HDF5 file. It defaults to "/X", but can be customized for other datasets such as "/coverage".

varColName

A string representing the regulation group within the HDF5 file. This is used to identify groups (or categories) within the dataset for computing the column-wise means.

Value

This function does not return any value but writes the computed mean values for each group into a new dataset in the HDF5 file.

Examples

## input data
suppressMessages(library(HDF5Array))
outputDir <- tempdir()
toy <- system.file('extdata', 'toy.h5', package = 'HDF5Array')
X <- HDF5Array(toy, 'M1')
X <- round(X * 1000, 0)
nRows <- dim(X)[1]
startN <- seq(from = 10000, by = 1000, length.out = nRows)
endN <- startN + 1000

## meta file
marker_name <- paste0('chr1_', startN, '_', endN)
sample_name <- paste0('sample_', 1:dim(X)[2])
group <- c(rep('case', 75), rep('ctrl', 75))
vargroup <- c(rep('intron', 3000), rep('exon', 3000), rep('utr', 4000))
metH5 <- file.path(outputDir, 'methylTracer_obj_test.h5')
unlink(metH5)

## build metH5
rhdf5::h5createFile(metH5)
rhdf5::h5createGroup(metH5, 'obs')
rhdf5::h5createGroup(metH5, 'var')
HDF5Array::writeHDF5Array(X, metH5, 'X')
#> <10000 x 150> HDF5Matrix object of type "double":
#>            [,1]   [,2]   [,3] ... [,149] [,150]
#>     [1,]    440     23    117   .    668    986
#>     [2,]    189    843    122   .    289    275
#>     [3,]    992    229    210   .    479    709
#>     [4,]    526    154    433   .    872    587
#>     [5,]    624    276    540   .    308    293
#>      ...      .      .      .   .      .      .
#>  [9996,]    692    493    692   .    505    175
#>  [9997,]     16    642    706   .    808     94
#>  [9998,]    519    786    515   .    763    184
#>  [9999,]    291    217    549   .    145    975
#> [10000,]    875    347     67   .    714    221
rhdf5::h5write(marker_name, metH5, 'var/marker_name')
rhdf5::h5write(vargroup, metH5, 'var/vargroup')
rhdf5::h5write(sample_name, metH5, 'obs/sample_name')
rhdf5::h5write(group, metH5, 'obs/group')

## build methylTracer object
met <- build_met_obj(metH5,
    sample_name = 'sample_name',
    marker_name = 'marker_name'
)

## compute var group values
computeVar(met = met, dataset = '/X', varColName = 'vargroup')
#> Saving results in /obs/mean_X_intron
#> Saving results in /obs/mean_X_exon
#> Saving results in /obs/mean_X_utr