Retrieve observations from a methylTracer object

Extract selected columns (and optionally selected rows) from the "/obs" group of the HDF5 file underlying a methylTracer object. The result is returned as a tibble-like data frame with one row per cell/sample.

metObs(met, indexList = NULL)

Arguments

met

A methylTracer object containing methylation data. The path to the HDF5 file is taken from met@seed@filepath.

indexList

Optional list of length 2 specifying which rows and columns of the "/obs" group to read. It should be of the form list(row_index, col_index), where:

row_index is an integer vector of row indices (1-based). If NULL, all rows are returned.

col_index is an integer vector of column indices (1-based) corresponding to datasets in "/obs". If NULL, all columns are returned.

The default (NULL) returns all rows and all columns in the "/obs" group.

Value

A tibble-like data frame (created via dplyr::bind_cols()) containing the selected observation-level metadata. Each column corresponds to one dataset in "/obs" (e.g. sample_name, group, etc.). If no rows are selected or "/obs" is empty, an empty data frame is returned and a message is printed.

Examples

## build a small HDF5 file with obs/var and create a methylTracer object
suppressMessages(library(HDF5Array))
outputDir <- tempdir()
toy <- system.file("extdata", "toy.h5", package = "HDF5Array")
X <- HDF5Array(toy, "M1")

nRows <- dim(X)[1]
startN <- seq(from = 10000, by = 1000, length.out = nRows)
endN <- startN + 1000

marker_name <- paste0("chr1_", startN, "_", endN)
sample_name <- paste0("sample_", seq_len(dim(X)[2]))
group <- c(rep("case", 75), rep("ctrl", 75))
vargroup <- c(rep("intron", 3000), rep("exon", 3000), rep("utr", 4000))

metH5 <- file.path(outputDir, "methylTracer_obj_test.h5")
unlink(metH5)

rhdf5::h5createFile(metH5)
rhdf5::h5createGroup(metH5, "obs")
rhdf5::h5createGroup(metH5, "var")
HDF5Array::writeHDF5Array(X, metH5, "X")
#> <10000 x 150> HDF5Matrix object of type "double":
#>                [,1]       [,2]       [,3] ...     [,149]     [,150]
#>     [1,] 0.44015009 0.02303051 0.11688992   .  0.6684808  0.9859835
#>     [2,] 0.18883376 0.84313244 0.12153315   .  0.2886549  0.2750908
#>     [3,] 0.99232698 0.22924928 0.20998143   .  0.4786133  0.7087926
#>     [4,] 0.52631346 0.15438076 0.43283016   .  0.8715839  0.5870442
#>     [5,] 0.62369062 0.27622849 0.54027815   .  0.3082406  0.2929455
#>      ...          .          .          .   .          .          .
#>  [9996,] 0.69183231 0.49322320 0.69210429   . 0.50506635 0.17500671
#>  [9997,] 0.01577389 0.64192487 0.70604724   . 0.80835509 0.09372537
#>  [9998,] 0.51859423 0.78645107 0.51546450   . 0.76275510 0.18370904
#>  [9999,] 0.29122618 0.21718068 0.54856003   . 0.14534205 0.97475079
#> [10000,] 0.87539984 0.34665351 0.06661729   . 0.71398136 0.22140136
rhdf5::h5write(marker_name, metH5, "var/marker_name")
rhdf5::h5write(vargroup, metH5, "var/vargroup")
rhdf5::h5write(sample_name, metH5, "obs/sample_name")
rhdf5::h5write(group, metH5, "obs/group")

met <- build_met_obj(
  metH5,
  sample_name = "sample_name",
  marker_name = "marker_name"
)

## all obs columns, all rows
obs_all <- metObs(met)

## only first 5 rows and the first 2 obs columns
obs_subset <- metObs(met, indexList = list(1:5, 1:2))